Migrate nodes to torch.nn.Module
- Added tests for the transfers
- All buffers that are registered and handled appropriately by the
unerlying torch.nn.Module object
