+++ bindsnet/network/utils.py
@@ -1,0 +1,1 @@
+import torch.nn as nn
