+++ test/network/test_nodes.py
@@ -1,6 +1,7 @@
 import torch
 
 from bindsnet.network.nodes import (
+    Nodes,
     Input,
     McCullochPitts,
     IFNodes,
@@ -50,3 +51,33 @@ class TestNodes:
                 assert layer.tc_decay == 1.5e3
                 assert (layer.s.float() == torch.zeros(n)).all()
                 assert (layer.v == layer.rest * torch.ones(n)).all()
+
+    def test_transfer(self):
+        for nodes in Nodes.__subclasses__():
+            layer = nodes(10)
+
+            layer.to(torch.device('cuda:0'))
+
+            layer_tensors = [k for k, v in layer.state_dict().items() if
+                    isinstance(v, torch.Tensor)]
+
+            tensor_devs = [getattr(layer,k).device for k in layer_tensors]
+
+            print("State dict in {} : {}".format(nodes,
+                layer.state_dict().keys()))
+            print("__dict__ in {} : {}".format(nodes,
+                layer.__dict__.keys()))
+            print("Tensors in {} : {}".format(nodes, layer_tensors))
+            print("Tensor devices {}".format(list(zip(layer_tensors,
+                tensor_devs))))
+
+            for d in tensor_devs:
+                print(d, d==torch.device('cuda:0'))
+                assert d == torch.device('cuda:0')
+
+
+if __name__ == "__main__":
+    tester = TestNodes()
+
+    tester.test_init()
+    tester.test_transfer()
